Move WHAT_141 and WHAT_153 to api_tests/courses package, move BaseTest to api_tests package.
