+++ src/main/java/api_tests/BaseTest.java
@@ -1,4 +1,4 @@
-package test.courses;
+package api_tests;
 
 import io.restassured.http.ContentType;
 import io.restassured.response.Response;
