+++ src/main/java/api_tests/courses/WHAT_141.java
@@ -1,5 +1,6 @@
-package test.courses;
+package api_tests.courses;
 
+import api_tests.BaseTest;
 import org.testng.annotations.Test;
 
 import static io.restassured.RestAssured.given;
